feat: use resize plugin to resize incoming frame in our plugin code
This is a WIP: The JS side of our plugin would take the incoming frame and resize it with vision-camera-resize-plugin. The resized data is then passed to our frame processor as additional param as currently there is no easy/efficient way to replace the original frame with a new resized frame.
The native side on Android would then need to use the resized data as input for the TF model directly instead of relying on Bitmap functions to resize, rescale and creating ByteBuffers on the Java side of things.
I tried to use the parameter on the native Android side but could not get to a meaningful space within the time I gave myself for this.